movie/requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
coverage==7.10.2
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
Faker==37.5.3
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-json-logger==3.3.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
